refactor: move constant string at product
