+++ package.json
@@ -18,8 +18,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.0",
-    "redux": "^4.2.1",
-    "redux-thunk": "^2.4.2"
+    "redux": "^4.2.1"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.195",
