+++ src/assets/product.ko.ts
@@ -1,5 +1,8 @@
 const PRODUCT = {
   NO_PRODUCT: "제품이 없습니다.",
+  FREE_SHIPPING: "무료배송",
+  SPECIAL_PRICE: "특가",
+  FROM_PRICE: "원 부터",
 };
 
 export { PRODUCT };
